requirements.txt
opencv-python
ffmpeg-python
scipy
tensorboard 
requests==2.27.1
pillow
wandb
matplotlib
timm
numpy
prettytable
torchsummary
torchinfo
seaborn
pandas
flow-vis
sailenv
tdw==1.9.3.0
scikit-learn
scikit-image
umap
